feat: add PHP views and htmx navigation
