+++ server/index.php
@@ -3,7 +3,7 @@
 require_once __DIR__.'/config/config.php';
 
 $route = trim(parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH), '/') ?: 'home';
-$titleMap = ['home'=>'PWA Template','login'=>'Login','register'=>'Register','users'=>'Users'];
+ $titleMap = ['home'=>'PWA Template','login'=>'Login','register'=>'Register','users'=>'Users','about'=>'About'];
 $title = $titleMap[$route] ?? ucfirst($route);
 
 $viewPath = __DIR__ . "/views/{$route}.php";
