+++ server/views/about.php
@@ -1,0 +1,4 @@
+<div data-island="about">
+    <h1>About</h1>
+    <p>About this app.</p>
+</div>
