+++ server/views/users.php
@@ -1,0 +1,5 @@
+<div data-island="users">
+    <h1>Users</h1>
+    <ul id="users-list"></ul>
+    <div id="users-message"></div>
+</div>
