+++ src/islands/about.ts
@@ -1,0 +1,3 @@
+export default function init() {
+    // No dynamic behavior needed for the about page
+}
