--- src/routes/about.ts
@@ -1,6 +1,0 @@
-export default function init() {
-    const content = document.getElementById('content');
-    if (content) {
-        content.innerHTML = '<h1>About</h1><p>About this app.</p>';
-    }
-}
